+++ compiler/src/policybuilder.rs
@@ -5,8 +5,7 @@ use std::collections::HashMap;
 use std::env;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
-use crate::config::IcmpFlowType;
-use crate::crypto::load_asn1data_from_pem;
+use crate::protocols::IcmpFlowType;
 use crate::errors::CompilationError;
 use crate::fabric::{Fabric, FabricService, ServiceType};
 use crate::polio;
