+++ compiler/src/protocols.rs
@@ -45,3 +45,49 @@ impl IanaProtocol {
         !self.is_icmp()
     }
 }
+
+#[derive(Debug, Clone)]
+pub struct Protocol {
+    pub protocol: IanaProtocol,
+    pub port: Option<String>, // TODO: using string for now but needs to be a "port spec"
+    pub icmp: Option<IcmpFlowType>,
+}
+
+/// Part of a Protocol description.
+#[derive(Debug, Clone, PartialEq)]
+pub enum IcmpFlowType {
+    RequestResponse(u8, u8),
+    OneShot(Vec<u8>),
+}
+
+impl Protocol {
+    pub fn to_endpoint_str(&self) -> String {
+        let mut s = String::new();
+        let protname = self.protocol.to_string().to_uppercase();
+        if self.protocol.is_icmp() {
+            if let Some(ref icmp) = self.icmp {
+                match icmp {
+                    IcmpFlowType::RequestResponse(req, resp) => {
+                        s.push_str(&format!("{}/{}", protname, req));
+                        s.push_str(&format!(",{}/{}", protname, resp));
+                    }
+                    IcmpFlowType::OneShot(ref codes) => {
+                        for c in codes {
+                            s.push_str(&format!("{}/{}", protname, c));
+                        }
+                    }
+                }
+            } else {
+                s.push_str(&format!("{}/0", protname));
+            }
+        } else {
+            s.push_str(&format!("{}/", protname));
+            if let Some(ref port) = self.port {
+                s.push_str(port);
+            } else {
+                s.push_str("0");
+            }
+        }
+        s
+    }
+}
